Partitioning the disk (and setting the RAID tag on the partition) BEFORE adding to the RAID array.
The partitions are added to the array, and not the disk itself. This change was suggested by the Azure support personnel, to help performance.

Also, using GPT for all partition tables. this is because, by default, we will be adding 1023 GB data disks, and if we add even 4 of them, the striped total size would be 4 TB. MBR partitions have a limit of 2 TB.